+++ exporter/opentelemetry-exporter-prometheus/tests/test_prometheus_exporter.py
@@ -36,8 +36,8 @@ class TestPrometheusMetricExporter(unittest.TestCase):
     def setUp(self):
         set_meter_provider(metrics.MeterProvider())
         self._meter = get_meter_provider().get_meter(__name__)
-        self._test_metric = self._meter.create_metric(
-            "testname", "testdesc", "unit", int, metrics.Counter,
+        self._test_metric = self._meter.create_counter(
+            "testname", "testdesc", "unit", int,
         )
         labels = {"environment": "staging"}
         self._labels_key = get_dict_as_key(labels)
@@ -80,8 +80,8 @@ class TestPrometheusMetricExporter(unittest.TestCase):
 
     def test_min_max_sum_aggregator_to_prometheus(self):
         meter = get_meter_provider().get_meter(__name__)
-        metric = meter.create_metric(
-            "test@name", "testdesc", "unit", int, metrics.ValueRecorder, []
+        metric = meter.create_valuerecorder(
+            "test@name", "testdesc", "unit", int, []
         )
         labels = {}
         key_labels = get_dict_as_key(labels)
@@ -101,9 +101,7 @@ class TestPrometheusMetricExporter(unittest.TestCase):
 
     def test_counter_to_prometheus(self):
         meter = get_meter_provider().get_meter(__name__)
-        metric = meter.create_metric(
-            "test@name", "testdesc", "unit", int, metrics.Counter,
-        )
+        metric = meter.create_counter("test@name", "testdesc", "unit", int,)
         labels = {"environment@": "staging", "os": "Windows"}
         key_labels = get_dict_as_key(labels)
         aggregator = SumAggregator()
@@ -134,9 +132,7 @@ class TestPrometheusMetricExporter(unittest.TestCase):
 
     def test_invalid_metric(self):
         meter = get_meter_provider().get_meter(__name__)
-        metric = meter.create_metric(
-            "tesname", "testdesc", "unit", int, StubMetric
-        )
+        metric = StubMetric("tesname", "testdesc", "unit", int, meter)
         labels = {"environment": "staging"}
         key_labels = get_dict_as_key(labels)
         record = MetricRecord(