+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+- Make `SpanProcessor.on_start` accept parent Context
+  ([#1251](https://github.com/open-telemetry/opentelemetry-python/pull/1251))
+- Fix b3 propagator entrypoint
+  ([#1265](https://github.com/open-telemetry/opentelemetry-python/pull/1265))
+- Allow None in sequence attributes values
+  ([#998](https://github.com/open-telemetry/opentelemetry-python/pull/998))
+- Samplers to accept parent Context
+  ([#1267](https://github.com/open-telemetry/opentelemetry-python/pull/1267))
+- Updating status codes to adhere to spec ([#1282](https://github.com/open-telemetry/opentelemetry-python/pull/1282))
+
 ## Version 0.14b0
 
 Released 2020-10-13